Réorganisation des vues dans `zds/tutorialv2/views`
