+++ zds/tutorialv2/urls/urls_publications.py
@@ -1,6 +1,6 @@
 from django.urls import re_path
 
-from zds.tutorialv2.views.published import ViewPublications
+from zds.tutorialv2.views.lists import ViewPublications
 from zds.tutorialv2.feeds import LastContentFeedRSS, LastContentFeedATOM
 
 urlpatterns = [
